move constants folder
